requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-dotenv
langchain-core
langchain-google-genai
langchain-tavily
langgraph
streamlit
httpx
sse-starlette
